NoteGropuingType moved
