+++ src/types/ChordMatch.ts
@@ -5,7 +5,7 @@ import { ChordDefinition } from "./ChordDefinition";
 import { ChordDisplayMode } from "./ChordDisplayMode";
 import { ActualIndex, ixInversion, InversionIndex } from "./IndexTypes";
 import { TWELVE } from "./NoteConstants";
-import { NoteGroupingType } from "./NoteGrouping";
+import { NoteGroupingType } from "./NoteGroupingTypes";
 import { NoteGroupingLibrary } from "./NoteGroupingLibrary";
 export class ChordMatch {
   constructor(
